requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
paddleocr
pillow
pytesseract
opencv-python

